+++ packages/implementations/react/src/atomic/Select/src/Select.tsx
@@ -1,6 +1,6 @@
 import React, { useEffect, useState } from "react";
 import { select } from "@rarui/styles";
-import { ArrowDownIcon } from "@rarui/icons";
+import { ArrowDownIcon, CloseIcon } from "@rarui/icons";
 import { Dropdown } from "@rarui-react/dropdown";
 import { Icon } from "@rarui-react/icon";
 import { Box } from "@rarui-react/box";
@@ -46,6 +46,11 @@ const Select: React.FC<SelectProps> = ({
     setSelectedOptions(selectedOptions.filter((item) => item !== option));
   };
 
+  const handleResetOptions = () => {
+    setOpen(true);
+    setSelectedOptions([]);
+  };
+
   useEffect(() => {
     if (value) {
       setSelectedOptions(value);
@@ -58,6 +63,8 @@ const Select: React.FC<SelectProps> = ({
     }
   }, [selectedOptions, onChange]);
 
+  const { className, style } = select.sprinkle(props);
+
   return (
     <Dropdown
       width="100%"
@@ -65,7 +72,10 @@ const Select: React.FC<SelectProps> = ({
       visible={open}
       onVisibility={handleOpen}
       content={
-        <Box display="flex" flexDirection="column" width="100%" gap="$3xs">
+        <div
+          className={[select.classnames.list, className].join(" ")}
+          style={style}
+        >
           {children}
           {options.map((option) => (
             <Dropdown.Item key={option.value} as="label" htmlFor={option.value}>
@@ -80,11 +90,11 @@ const Select: React.FC<SelectProps> = ({
               />
             </Dropdown.Item>
           ))}
-        </Box>
+        </div>
       }
     >
       <div className={select.classnames.select({ size })} {...props}>
-        <Box display="flex" flexWrap="wrap" gap="$3xs">
+        <Box display="flex" flex="1" gap="$3xs" overflowY="hidden">
           {selectedOptions.map((selectedOption) => (
             <Chip
               key={selectedOption.value}
@@ -98,7 +108,19 @@ const Select: React.FC<SelectProps> = ({
             </Chip>
           ))}
         </Box>
-        <Icon source={<ArrowDownIcon size="medium" />} />
+        <Box display="flex" alignItems="center" gap="$4xs" paddingLeft="$3xs">
+          {!!selectedOptions.length && (
+            <button
+              type="button"
+              className={select.classnames.close}
+              onClick={handleResetOptions}
+              aria-label="clear options"
+            >
+              <Icon source={<CloseIcon size="small" />} />
+            </button>
+          )}
+          <Icon source={<ArrowDownIcon size="medium" />} />
+        </Box>
       </div>
     </Dropdown>
   );