+++ packages/implementations/react/src/atomic/Select/src/select.stories.tsx
@@ -24,11 +24,19 @@ type Story = StoryObj<typeof Select>;
 
 export const basic: Story = {
   args: {
-    onChange: (options) => options,
+    onChange: (options) => console.log(options),
     value: [
       {
         label: "Item 1",
         value: "item-1",
+      },
+      {
+        label: "Item 2",
+        value: "item-2",
+      },
+      {
+        label: "Item 3",
+        value: "item-3",
       },
     ],
     options: [