+++ packages/implementations/react/src/composite/Tabs/src/components/Item/index.ts
@@ -1,0 +1,5 @@
+import { TabsItem } from "./Item";
+
+export { TabsItem } from "./Item";
+export type { TabsItemProps } from "./item.types";
+export default TabsItem;