+++ packages/implementations/react/src/composite/Tabs/src/index.ts
@@ -1,0 +1,5 @@
+import { Tabs } from "./Tabs";
+
+export { Tabs } from "./Tabs";
+export type { TabsProps } from "./tabs.types";
+export default Tabs;