+++ packages/styles/src/packages/atomic/select/rarui-select.css.ts
@@ -1,5 +1,10 @@
+import {
+  createRainbowSprinkles,
+  defineProperties as defineRainbowProperties,
+} from "rainbow-sprinkles";
 import { recipe } from "@vanilla-extract/recipes";
-import { varsThemeBase } from "../../../themes";
+import { style } from "@vanilla-extract/css";
+import { mediaQueries, varsThemeBase } from "../../../themes";
 
 const select = recipe({
   base: {
@@ -71,6 +76,73 @@ const select = recipe({
   },
 });
 
+const list = style({
+  display: "flex",
+  flexDirection: "column",
+  width: "100%",
+  overflowY: "scroll",
+  gap: varsThemeBase.spacing["3xs"],
+  "::-webkit-scrollbar": {
+    width: ".375rem",
+    background: "transparent",
+  },
+  "::-webkit-scrollbar-thumb": {
+    borderRadius: varsThemeBase.shape.border.radius["2xs"],
+    backgroundColor: varsThemeBase.colors.content.disabled,
+  },
+});
+
+const close = style({
+  display: "flex",
+  alignItems: "center",
+  justifyContent: "center",
+  cursor: "pointer",
+  background: "transparent",
+  color: "currentcolor",
+  transition: "all 150ms ease",
+  height: "1rem",
+  width: "1rem",
+  border: 0,
+  padding: 0,
+  borderRadius: varsThemeBase.shape.border.radius["2xs"],
+  ":focus-within": {
+    outline: "none",
+    boxShadow: varsThemeBase.utils["focus-ring"],
+  },
+  ":hover": {
+    color: varsThemeBase.colors.content.primary,
+  },
+});
+
 export const selectStyles = {
   select,
+  list,
+  close,
 };
+
+/* -------------------------------------------------------------------------------------------------
+ * Sprinkle
+ * -----------------------------------------------------------------------------------------------*/
+
+const defineProperties = defineRainbowProperties({
+  conditions: {
+    xs: {
+      "@media": mediaQueries.xs(),
+    },
+    md: {
+      "@media": mediaQueries.md(),
+    },
+    lg: {
+      "@media": mediaQueries.lg(),
+    },
+    xl: {
+      "@media": mediaQueries.xl(),
+    },
+  },
+  defaultCondition: "xs",
+  dynamicProperties: {
+    maxHeight: true,
+  },
+});
+
+export const selectSprinkle = createRainbowSprinkles(defineProperties);