+++ packages/styles/src/packages/atomic/select/rarui-select.types.ts
@@ -1,6 +1,9 @@
 import { RecipeVariants } from "@vanilla-extract/recipes";
 import { selectStyles } from "./rarui-select.css";
+import { StandardLonghandProperties } from "../../../index.types";
 
 export type SelectVariants = NonNullable<
   RecipeVariants<typeof selectStyles.select>
 >;
+
+export type SelectSprinkle = Pick<StandardLonghandProperties, "maxHeight">;