+++ packages/styles/src/packages/composite/tabs/index.ts
@@ -1,0 +1,7 @@
+import { tabsStyles } from "./rarui-tabs.css";
+
+export const tabs = {
+  classnames: tabsStyles,
+};
+
+export type { TabsVariants } from "./rarui-tabs.types";