+++ packages/implementations/react/src/atomic/Select/CHANGELOG.md
@@ -2,6 +2,17 @@
 
 Use a select box when a user needs to select one or more than one option from a list.
 
+## 2024-06-18 `1.0.1`
+
+#### 🎉 New features
+
+- Added button to clean the selected options. ([#58](https://git.rarolabs.com.br/frontend/rarui/-/merge_requests/58) by [@junior](https://git.rarolabs.com.br/junior))
+- Added `maxHeight` property to the component. ([#58](https://git.rarolabs.com.br/frontend/rarui/-/merge_requests/58) by [@junior](https://git.rarolabs.com.br/junior))
+
+#### 🐛 Bug fixes
+
+- Fixed function callback being called several times. ([#58](https://git.rarolabs.com.br/frontend/rarui/-/merge_requests/58) by [@junior](https://git.rarolabs.com.br/junior))
+
 ## 2024-06-12 `1.0.0`
 
 #### 🎉 New features